Move embeddings tests
